split intermediate handler
