+++ test/unit/Jamfile
@@ -74,7 +74,7 @@ run
         test/detail/connect_params_helpers.cpp
         test/detail/output_string.cpp
         test/detail/engine_impl.cpp
-        test/detail/async_helpers.cpp
+        test/detail/intermediate_handler.cpp
 
         test/detail/typing/meta_check_context.cpp
         test/detail/typing/pos_map.cpp
